@@ -1,16 +1,18 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-footer',
-  templateUrl: './footer.component.html',
-  styleUrls: ['./footer.component.css']
+    selector: 'app-footer',
+    templateUrl: './footer.component.html',
+    styleUrls: ['./footer.component.css']
 })
 export class FooterComponent implements OnInit {
-  test : Date = new Date();
-  
-  constructor() { }
 
-  ngOnInit() {
-  }
+    test: Date = new Date();
+
+    constructor() {
+    }
+
+    ngOnInit() {
+    }
 
 }
