@@ -4,18 +4,13 @@ import {RouterModule} from '@angular/router';
 import {FormsModule} from '@angular/forms';
 
 import {MatSelectModule} from '@angular/material/select';
-
-import {DashboardComponent} from './dashboard-component/dashboard.component';
-import {AreaGraphComponent} from './area-graph/area-graph.component';
-import {BarGraphComponent} from './bar-graph/bar-graph.component';
-import {GlucoseGraphComponent} from './glucose-graph/glucose-graph.component';
-import {LineGraphComponent} from './line-graph/line-graph.component';
-import {ScatterGraphComponent} from './scatter-graph/scatter-graph.component';
-import {SharedModule} from 'app/shared/shared.module';
-import {DashboardRoutingModule} from './dashboard-routing/dashboard-routing.module';
-import {DashboardService} from './services/dashboard.service';
 import {MatPaginatorModule} from "@angular/material";
 
+import {DashboardComponent} from './dashboard.component/dashboard.component';
+import {SharedModule} from 'app/shared/shared.module';
+import {DashboardRoutingModule} from './dashboard.routing/dashboard.routing.module';
+import {DashboardService} from './services/dashboard.service';
+import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
     imports: [
@@ -26,24 +21,13 @@ import {MatPaginatorModule} from "@angular/material";
         DashboardRoutingModule,
 
         MatSelectModule,
-        MatPaginatorModule
+        MatPaginatorModule,
+        TranslateModule
     ],
     declarations: [
-        DashboardComponent,
-        AreaGraphComponent,
-        BarGraphComponent,
-        GlucoseGraphComponent,
-        LineGraphComponent,
-        ScatterGraphComponent
+        DashboardComponent
     ],
-    providers: [DashboardService],
-    exports: [
-        AreaGraphComponent,
-        BarGraphComponent,
-        GlucoseGraphComponent,
-        LineGraphComponent,
-        ScatterGraphComponent
-    ],
+    providers: [DashboardService]
 })
 export class DashboardModule {
 }
