@@ -1,23 +1,18 @@
 export enum EvaluationStatus {
-    complete = "complete",
-    incomplete = "incomplete"
+    complete = 'complete',
+    incomplete = 'incomplete'
 }
 
-export interface ICounseling {
+export class Counseling {
     suggested: any;
     definitive: any
 }
 
-export interface IEvaluation {
-    id: string;
-    created_at: string;
-
-}
-
-export class Evaluation implements IEvaluation {
+export class Evaluation {
 
     id: string;
     created_at: string;
 
-    constructor() { }
+    constructor() {
+    }
 }
