@@ -1,75 +1,52 @@
-import {Component, OnInit, Input, SimpleChanges, OnChanges} from '@angular/core';
-import {DatePipe} from '@angular/common';
+import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import { DatePipe } from '@angular/common';
 
-import * as _ from 'lodash';
+import { TranslateService } from '@ngx-translate/core';
 
-import {IMeasurement} from '../models/measurement';
-import {GraphService} from 'app/shared/shared-services/graph.service';
+import { Measurement, MeasurementType } from '../models/measurement';
+import { Weight } from '../models/weight';
+import { MeasurementService } from '../services/measurement.service';
 
 @Component({
     selector: 'height',
     templateUrl: './height.component.html',
-    styleUrls: ['./height.component.scss']
+    styleUrls: ['../shared.style/shared.styles.scss']
 })
 export class HeightComponent implements OnInit, OnChanges {
-
-    @Input() data: Array<IMeasurement>;
-
-    lastData: IMeasurement;
-
-    option = {
-        tooltip: {
-            trigger: 'axis',
-            formatter: "Altura: {c} cm <br> Data: {b}"
-        },
-        xAxis: {
-            type: 'category',
-            data: []
-        },
-        yAxis: {
-            type: 'value',
-            axisLabel: {
-                formatter: '{value} cm'
-            }
-        },
-        dataZoom: [
-            {
-                type: 'slider'
-            }
-        ],
-        series: [
-            {
-                name: 'Altura',
-                type: 'line',
-                step: 'start',
-                data: []
-            }
-        ]
-    };
-
+    @Input() data: Array<Measurement>;
+    @Input() filter_visibility: boolean;
+    @Input() patientId: string;
+    lastData: Measurement;
+    options: any;
+    showSpinner: boolean;
+    echartsInstance: any;
 
     constructor(
         private datePipe: DatePipe,
-        private graphService: GraphService
+        private measurementService: MeasurementService,
+        private translateService: TranslateService
     ) {
-        this.data = new Array<IMeasurement>();
-
+        this.data = new Array<Measurement>();
+        this.filter_visibility = false;
+        this.patientId = '';
+        this.showSpinner = false;
     }
 
-    ngOnInit() {
+    ngOnInit(): void {
+        this.loadGraph();
+    }
 
+    onChartInit(event) {
+        this.echartsInstance = event;
     }
 
     loadGraph() {
 
-        //Limpando o grafico
-        this.option.xAxis.data = [];
-        this.option.series[0].data = [];
-
-        this.data.forEach((element: IMeasurement) => {
-            this.option.xAxis.data.push(this.datePipe.transform(element.timestamp, "shortDate"));
-            this.option.series[0].data.push(element.value);
-        });
+        const height = this.translateService.instant('MEASUREMENTS.HEIGHT.HEIGHT');
+        const date = this.translateService.instant('SHARED.DATE-AND-HOUR');
+        const max = this.translateService.instant('MEASUREMENTS.MAX');
+        const min = this.translateService.instant('MEASUREMENTS.MIN');
+        const at = this.translateService.instant('SHARED.AT');
 
         if (this.data.length > 1) {
             this.lastData = this.data[this.data.length - 1];
@@ -77,9 +54,94 @@ export class HeightComponent implements OnInit, OnChanges {
             this.lastData = this.data[0];
         }
 
-        this.graphService.refreshGraph();
+        const xAxis = {
+            type: 'category',
+            data: []
+        };
+
+        const series = {
+            name: height,
+            type: 'line',
+            step: 'start',
+            data: [],
+            color: '#3F51B5',
+            markPoint: {
+                label: {
+                    fontSize: 10,
+                    formatter: function (params) {
+                        if (params.data.type === 'max') {
+                            return max;
+                        }
+                        if (params.data.type === 'min') {
+                            return min;
+                        }
+                    }
+                },
+                data: [
+                    { type: 'max' },
+                    { type: 'min' }
+                ]
+            }
+        };
+
+        this.data.forEach((element: Measurement) => {
+            xAxis.data.push(this.datePipe.transform(element.timestamp, 'shortDate'));
+            series.data.push({
+                value: element.value,
+                time: this.datePipe.transform(element.timestamp, 'mediumTime')
+            });
+        });
 
 
+        this.options = {
+            tooltip: {
+                trigger: 'item',
+                formatter: function (params) {
+                    return `${height}: ${params.data.value} cm <br> ${date}: <br> ${params.name} ${at} ${params.data.time}`;
+                }
+            },
+            xAxis: xAxis,
+            yAxis: {
+                type: 'value',
+                axisLabel: {
+                    formatter: '{value} cm'
+                }
+            },
+            dataZoom: [
+                {
+                    type: 'slider'
+                }
+            ],
+            series: series
+        };
+
+
+    }
+
+    applyFilter(filter: { start_at: string, end_at: string, period: string }) {
+        this.showSpinner = true;
+        this.measurementService.getAllByUserAndType(this.patientId, MeasurementType.height, null, null, filter)
+            .then((measurements: Array<any>) => {
+                this.data = measurements;
+                this.showSpinner = false;
+                this.updateGraph(measurements);
+            })
+            .catch();
+    }
+
+    updateGraph(measurements: Array<any>): void {
+        // clean weightGraph
+        this.options.xAxis.data = [];
+        this.options.series.data = [];
+
+        measurements.forEach((element: Weight) => {
+            this.options.xAxis.data.push(this.datePipe.transform(element.timestamp, 'shortDate'));
+            this.options.series.data.push({
+                value: element.value,
+                time: this.datePipe.transform(element.timestamp, 'mediumTime')
+            });
+        });
+        this.echartsInstance.setOption(this.options);
     }
 
     ngOnChanges(changes: SimpleChanges) {
