@@ -1,14 +1,28 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+
 import { MatPaginatorIntl } from '@angular/material';
-import { getConfigPaginator } from './config-matpaginator';
+import { TranslateService } from '@ngx-translate/core';
+
+import { PaginatorIntlService } from './config.matpaginator';
+
 @NgModule({
-  declarations: [],
-  imports: [
-    CommonModule
-  ],
-  providers: [
-    { provide: MatPaginatorIntl, useValue: getConfigPaginator() }
-  ]
+    declarations: [],
+    imports: [
+        CommonModule
+    ],
+    providers: [
+        PaginatorIntlService,
+        {
+            provide: MatPaginatorIntl,
+            useFactory: (translate) => {
+                const service = new PaginatorIntlService();
+                service.injectTranslateService(translate);
+                return service;
+            },
+            deps: [TranslateService]
+        }
+    ]
 })
-export class ModulesModule { }
+export class ModulesModule {
+}
