@@ -8,24 +8,25 @@ import { MatDatepickerModule } from '@angular/material/datepicker';
 import { MatSelectModule } from '@angular/material/select';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatTabsModule } from '@angular/material';
+import { TranslateModule } from '@ngx-translate/core';
 
-import { PatientFormComponent } from './patient-form/patient-form.component';
-import { PatientTableComponent } from './patient-table/patient-table.component';
-import { PatientComponentComponent } from './patient-component/patient-component.component';
+import { PatientFormComponent } from './patient.form/patient.form.component';
+import { PatientTableComponent } from './patient.table/patient.table.component';
+import { PatientComponent } from './patient.component/patient.component';
 
-import { PatientRoutingModule } from './patient-routing/patient-routing.module';
+import { PatientRoutingModule } from './patient.routing/patient.routing.module';
 import { SharedModule } from 'app/shared/shared.module';
-import { ListPilotstudiesComponent } from './list-pilotstudies/list-pilotstudies.component';
-import { ViewHabitsComponent } from './view-habits/view-habits.component';
+import { ListPilotstudiesComponent } from './list.pilotstudies/list.pilotstudies.component';
+import { ViewHabitsComponent } from './view.habits/view.habits.component';
 import { HabitsModule } from '../habits/habits.module';
 import { MeasurementModule } from '../measurement/measurement.module';
-import { PatientManagerComponent } from './patient-manager/patient-manager.component';
+import { PatientManagerComponent } from './patient.manager/patient.manager.component';
 
 @NgModule({
   declarations: [
     PatientFormComponent,
     PatientTableComponent,
-    PatientComponentComponent,
+    PatientComponent,
     ListPilotstudiesComponent,
     ViewHabitsComponent,
     PatientManagerComponent
@@ -42,6 +43,7 @@ import { PatientManagerComponent } from './patient-manager/patient-manager.compo
     MatNativeDateModule,
     MatFormFieldModule,
     MatTabsModule,
+    TranslateModule,
 
     PatientRoutingModule,
     HabitsModule,
