@@ -1,25 +1,30 @@
-import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
-import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from '@angular/common/http';
-import { environment } from 'environments/environment';
+import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
+
+import { Observable } from 'rxjs';
+
+import { LocalStorageService } from '../../shared/shared.services/localstorage.service';
 
 @Injectable()
 export class TokenInterceptor implements HttpInterceptor {
-  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    const requestUrl: Array<any> = request.url.split('/');
-    const apiUrl: Array<any> = environment.api_url.split('/');
-    const token = localStorage.getItem('token');
-    
-    if (token) {//&& (requestUrl[2] === apiUrl[2])
-      const newRequest = request.clone({ setHeaders: {
-        'Authorization': `Bearer ${token}`,
-        'Content-Type': 'application/json'
-      } 
-      });
-      return next.handle(newRequest);
-    } else {
-      return next.handle(request);
+    constructor(
+        private localStorageService: LocalStorageService) {
     }
 
-  }
+    intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+        const token = this.localStorageService.getItem('token');
+
+        if (token) {
+            const newRequest = request.clone({
+                setHeaders: {
+                    'Authorization': `Bearer ${token}`,
+                    'Content-Type': 'application/json'
+                }
+            });
+            return next.handle(newRequest);
+        } else {
+            return next.handle(request);
+        }
+
+    }
 }
