@@ -1,5 +1,11 @@
 /* SystemJS module definition */
 declare var module: NodeModule;
+
 interface NodeModule {
-  id: string;
+    id: string;
+}
+
+declare module '*.html' {
+    const value: string;
+    export default value
 }
